fix: swapped all int32_t to int16_t

+++ src/include/task.h
@@ -30,28 +30,28 @@ typedef enum {
 typedef struct {
   char name[MAX_STR_LEN + 1]; /* task name                      */
   
-  int32_t (*addr)();         /* first instruction address      */
+  int16_t (*addr)();         /* first instruction address      */
   
   task_type type;             /* task type                      */
   task_state state;           /* task state                     */
   task_crit criticality;      /* task criticality               */
   
-  long dline;                 /* task absolute deadline         */
+  int32_t dline;                 /* task absolute deadline         */
   
-  int32_t period;            /* task period                    */
-  int32_t priority;          /* task priority                  */
-  int32_t wcet;              /* task worst-case execution time */
+  int16_t period;            /* task period                    */
+  int16_t priority;          /* task priority                  */
+  int16_t wcet;              /* task worst-case execution time */
   
   float utilf;                /* task utilization factor        */
   
-  int32_t *ctx;               /* pointer to the task's context  */
+  int16_t *ctx;               /* pointer to the task's context  */
   
-  int32_t next;              /* pointer to the next TCB        */
-  int32_t prev;              /* pointer to the previous TCB    */
+  int16_t next;              /* pointer to the next TCB        */
+  int16_t prev;              /* pointer to the previous TCB    */
 } TCB;
 
-int32_t t1(void) { while(TRUE) { Serial.print("AaaaaaA"); } return 1; }
-int32_t t2(void) { while(TRUE) { Serial.print("BbbbbbbbbB"); } return 1; }
-int32_t t3(void) { while(TRUE) { Serial.print("CcC"); } return 1; }
+int16_t t1(void) { while(TRUE) { Serial.print("AaaaaaA"); } return 1; }
+int16_t t2(void) { while(TRUE) { Serial.print("BbbbbbbbbB"); } return 1; }
+int16_t t3(void) { while(TRUE) { Serial.print("CcC"); } return 1; }
 
 #endif // TASK_H
